requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachier==1.4.2
chardet==3.0.4
idna==2.9
multidict==4.7.5
pathtools==0.1.2
portalocker==1.6.0
slackclient==2.5.0
watchdog==0.10.2
yarl==1.4.2
pandas==1.0.3